edit Couchbase license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit creates  LICENSE file containing the full text
of the Apache 2.0 license. This allows Licensee to successfully
identify the license type of Couchbase Server's codebase as Apache 2.0.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>